Refactor out utilities functions
